Back out 79684dc64173 (bug 898936) for 10.7 debug M4 crashes

+++ js/src/jit/PerfSpewer.cpp
@@ -149,7 +149,7 @@ PerfSpewer::writeProfile(JSScript *script,
         if (size > 0) {
             fprintf(fp_,
                     "%lx %lx %s:%d: Func%02d\n",
-                    reinterpret_cast<unsigned long>(code->raw()),
+                    reinterpret_cast<uintptr_t>(code->raw()),
                     size,
                     script->filename(),
                     script->lineno,
@@ -170,8 +170,7 @@ PerfSpewer::writeProfile(JSScript *script,
             if (cur < blockStart) {
                 fprintf(fp_,
                         "%lx %lx %s:%d: Func%02d-Block?\n",
-                        static_cast<unsigned long>(cur),
-                        static_cast<unsigned long>(blockStart - cur),
+                        cur, blockStart - cur,
                         script->filename(), script->lineno,
                         thisFunctionIndex);
             }
@@ -182,7 +181,7 @@ PerfSpewer::writeProfile(JSScript *script,
             if (size > 0) {
                 fprintf(fp_,
                         "%lx %lx %s:%d:%d: Func%02d-Block%d\n",
-                        static_cast<unsigned long>(blockStart), size,
+                        blockStart, size,
                         r.filename, r.lineNumber, r.columnNumber,
                         thisFunctionIndex, r.id);
             }
@@ -194,8 +193,7 @@ PerfSpewer::writeProfile(JSScript *script,
         if (cur < funcEnd) {
             fprintf(fp_,
                     "%lx %lx %s:%d: Func%02d-OOL\n",
-                    static_cast<unsigned long>(cur),
-                    static_cast<unsigned long>(funcEnd - cur),
+                    cur, funcEnd - cur,
                     script->filename(), script->lineno,
                     thisFunctionIndex);
         }
